Linux 2.2.20pre2
o	Fix ip_decrease_ttl as per 2.4			(Dave Miller)
o	Fix tcp retransmit state bug			(Alexey Kuznetsov)
o	Fix a few obscure sparc tree bugs		(Dave Miller)
o	Fix fb /proc bug and OF fb name size bug	(Segher Boessenkool)
o	Fix complie with CONFIG_INTEL_RNG=y		(Andrzej Krzysztofowicz)
o	Fix rio driver when HZ!=100			(Andrzej Krzysztofowicz)
o	Stop 3c509 grabbing other EISA boards		(Andrzej Krzysztofowicz)
o	Remove surplus defines for root= names		(Andrzej Krzysztofowicz)
o	Revert pre1 APIC change

+++ include/net/ip.h
@@ -170,7 +170,7 @@ extern __inline__ void ip_send(struct sk_buff *skb)
 extern __inline__
 int ip_decrease_ttl(struct iphdr *iph)
 {
-	u16 check = iph->check;
+	u32 check = iph->check;
 	check += __constant_htons(0x0100);
 	iph->check = check + ((check>=0xFFFF) ? 1 : 0);
 	return --iph->ttl;
